refactor(tui): split app.py into dispatch / runner / keybindings modules
tui/app.py era un monolito de 906+ LOC que mezclaba:

  - Composicion del widget tree
  - Carga de instances.json
  - BINDINGS y los action_* handlers
  - Pipeline de dispatch (action -> modal -> worker)
  - Ejecucion de subprocesses (streamed + interactive)
  - Log sink (_log/_log_bulk)
  - Helpers de arg (arg_to_field, arg_defaults, warn_unknown_modules)

Lo partimos en 3 modulos para que cada uno sea legible y testeable
en aislamiento, y para que cambios en una concern no toquen las
otras:

- `tui/dispatch.py` (DispatchMixin): el pipeline action -> modal ->
  worker. Incluye _dispatch, _update_picker_async, _dispatch_fallback,
  _arg_to_field, _arg_defaults, _on_modal_result, _on_picker_result,
  _on_picker_result_with_db, _warn_unknown_modules, _execute,
  _execute_all, _execute_one, _is_update_with_modules,
  _run_interactive.

- `tui/keybindings.py` (KeybindingsMixin): BINDINGS + todos los
  action_* handlers. Incluye _save_instances_json_async,
  _save_instances_json (shim sync), _rebuild_richlog_from_up,
  _rebuild_richlog_async, _toggle_log_level, action_refresh.

- `tui/runner.py`: ahora ademas de stream_command tiene
  run_interactive, que reemplaza el viejo `asyncio.to_thread(
  subprocess.run, ...)` por una coroutine asyncio pura.

- `tui/app.py` queda en 455 LOC, 10 metodos: solo composition
  (__init__, compose, on_mount), event wiring (on_list_view_*),
  data loading (refresh_instances, refresh_actions,
  _scan_instance_modules), log sink (_log/_log_bulk) y el
  _run_streamed (que esta atado al estado de UI per-instance:
  progress widget, RichLog, throttle, asi que moverlo seria
  pasar demasiados callbacks).

Por que mixin y no free functions: la mayoria de los metodos
necesitan self (self._log, self._raw_config, self.config, ...).
Un mixin da la sintaxis natural de metodo de instancia sin el
`app.` repetido que tendrian las free functions, y Python
resuelve el MRO trivialmente con
`class DockerOdooApp(DispatchMixin, KeybindingsMixin, App)`.

tui.app re-exporta los helpers puros `_scan_instance_modules_pure`
y `_write_instances_json` (necesarios para que dispatch y
keybindings puedan llamarlos via to_thread sin tocar el DOM).

Tests: 45 OK, todos sin modificacion. `DockerOdooApp` se sigue
importando desde `tui.app` como antes.
